backend/requirements.txt
anyio==4.2.0
appdirs==1.4.4
asgiref==3.7.2
attrs==23.2.0
backoff==2.2.1
boilerpy3==1.0.7
cattrs==23.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.1
django-cors-headers==4.3.1
django-import-export==3.3.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
et-xmlfile==1.1.0
Events==0.5
exceptiongroup==1.2.0
faiss-cpu==1.7.2
farm-haystack==1.24.1
filelock==3.13.1
fsspec==2024.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.6
inflect==7.0.0
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lazy-imports==0.3.1
MarkupPy==1.14
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
nltk==3.8.1
num2words==0.5.13
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pandas==2.2.0
phonenumbers==8.13.27
pillow==10.2.0
platformdirs==4.2.0
posthog==3.4.1
prompthub-py==4.0.0
psycopg==3.1.16
psycopg-binary==3.1.16
psycopg2==2.9.9
pydantic==1.10.14
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
quantulum3==0.9.0
rank-bm25==0.2.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-cache==0.9.8
rpds-py==0.18.0
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
sentence-transformers==2.3.1
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.51
sqlparse==0.4.4
sseclient-py==1.8.0
sympy==1.12
tablib==3.5.0
tenacity==8.2.3
threadpoolctl==3.3.0
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.0
tqdm==4.66.2
transformers==4.36.2
typing_extensions==4.9.0
tzdata==2023.4
url-normalize==1.4.3
urllib3==2.2.0
xlrd==2.0.1
xlwt==1.3.0
